+++ polaris-ui/src/app/common/presentation/components/LinkButton.module.scss
@@ -13,14 +13,9 @@
     margin: 0;
     border: 0.125rem solid;
     border-color: #ffdd00;
-    outline: 0.1875rem solid transparent;
-    box-shadow: inset 0 0 0 0.0625rem #ffdd00;
-  }
-
-  &:focus:not(:active):not(:hover) {
-    border-color: #ffdd00;
     color: #0b0c0c;
     background-color: #ffdd00;
     box-shadow: 0 0.125rem 0 #0b0c0c;
+    outline: 0.1875rem solid transparent;
   }
 }
