+++ polaris-ui/src/app/common/presentation/components/tabs/TabButtons.tsx
@@ -1,0 +1,148 @@
+import { useEffect, useState, useRef } from "react";
+import classes from "./Tabs.module.scss";
+import { ItemProps } from "./types";
+import { ReactComponent as CloseIcon } from "../../svgs/closeIconBold.svg";
+import { ReactComponent as ArrowLeft } from "../../svgs/arrowLeft.svg";
+import { ReactComponent as ArrowRight } from "../../svgs/arrowRight.svg";
+import { LinkButton } from "../../../../common/presentation/components/LinkButton";
+export type TabButtonProps = {
+  items: ItemProps[];
+  activeTabIndex: number;
+  handleTabSelection: (documentId: string) => void;
+  handleCloseTab: () => void;
+  handleUnLockDocuments: (documentIds: string[]) => void;
+};
+
+const ARROW_KEY_SHIFTS = {
+  ArrowLeft: -1,
+  ArrowUp: -1,
+  ArrowRight: 1,
+  ArrowDown: 1,
+};
+
+const TabButtons: React.FC<TabButtonProps> = ({
+  items,
+  activeTabIndex,
+  handleTabSelection,
+  handleCloseTab,
+}) => {
+  const activeTabRef = useRef<HTMLButtonElement>(null);
+
+  useEffect(() => {
+    activeTabRef.current?.focus();
+    activeTabRef.current?.parentElement?.scrollIntoView({
+      behavior: "smooth",
+      block: "nearest",
+    });
+  }, [activeTabIndex, items.length]);
+
+  const handleKeyPressOnTab: React.KeyboardEventHandler<HTMLButtonElement> = (
+    ev
+  ) => {
+    const typedKeyCode = ev.code as keyof typeof ARROW_KEY_SHIFTS;
+    const thisShift = ARROW_KEY_SHIFTS[typedKeyCode]; // -1, 1 or undefined
+
+    moveToNextOrPreviousTab(thisShift);
+    if (ev.code === "ArrowRight" || ev.code === "ArrowLeft") {
+      ev.preventDefault();
+    }
+  };
+
+  const moveToNextOrPreviousTab = (thisShift: number) => {
+    const shouldNavigate =
+      // must be a left or right key press command
+      !!thisShift &&
+      // can't go left on the first tab
+      !(activeTabIndex === 0 && thisShift === -1) &&
+      // can't go right on the last tab
+      !(activeTabIndex === items.length - 1 && thisShift === 1);
+
+    if (!shouldNavigate) {
+      return;
+    }
+
+    const nextTabIndex = activeTabIndex + thisShift;
+    const nextTabId = items[nextTabIndex].id;
+    handleTabSelection(nextTabId);
+  };
+  if (!items.length) {
+    return null;
+  }
+  return (
+    <div className={`${classes.tabsWrapper}`}>
+      <LinkButton
+        disabled={activeTabIndex === 0}
+        className={classes.tabNextButton}
+        dataTestId="btn-tab-next"
+        onClick={() => {
+          moveToNextOrPreviousTab(-1);
+        }}
+      >
+        <ArrowLeft />
+      </LinkButton>
+
+      <ul className={`${classes.tabsList}`} role="tablist">
+        {items.map((item, index) => {
+          const { id: itemId, label } = item;
+
+          return (
+            <li
+              className={`${
+                activeTabIndex === index
+                  ? classes.activeTab
+                  : classes.inactiveTab
+              } ${classes.tabListItem}`}
+              key={itemId}
+              data-testid={`tab-${index}`}
+              role="presentation"
+            >
+              <button
+                id={`tab_${index}`}
+                aria-controls={`panel-${index}`}
+                role="tab"
+                className={classes.tabButton}
+                data-testid={
+                  index === activeTabIndex ? "tab-active" : `btn-tab-${index}`
+                }
+                onClick={() => {
+                  if (itemId !== items[activeTabIndex].id) {
+                    handleTabSelection(itemId);
+                  }
+                }}
+                onKeyDown={handleKeyPressOnTab}
+                tabIndex={index === activeTabIndex ? 0 : -1}
+                ref={index === activeTabIndex ? activeTabRef : undefined}
+              >
+                <span className={classes.tabLabel}> {label}</span>
+              </button>
+
+              {activeTabIndex === index && (
+                <button
+                  className={classes.tabCloseButton}
+                  onClick={handleCloseTab}
+                  onKeyDown={handleKeyPressOnTab}
+                  data-testid="tab-remove"
+                  aria-label="close tab"
+                >
+                  <CloseIcon />
+                </button>
+              )}
+            </li>
+          );
+        })}
+      </ul>
+      <LinkButton
+        disabled={activeTabIndex === items.length - 1}
+        className={classes.tabNextButton}
+        dataTestId="btn-tab-next"
+        onClick={() => {
+          moveToNextOrPreviousTab(1);
+        }}
+      >
+        <ArrowRight />
+      </LinkButton>
+    </div>
+  );
+};
+
+export default TabButtons;
