+++ polaris-ui/src/app/common/presentation/components/tabs/Tabs.tsx
@@ -1,17 +1,10 @@
-import { useEffect, useState, useRef } from "react";
+import { useState } from "react";
 import { CommonTabsProps } from "./types";
 import { useLastFocus } from "../../../hooks/useLastFocus";
 import { Modal } from "../../../../common/presentation/components/Modal";
 import { NavigationAwayAlertContent } from "../../../../features/cases/presentation/case-details/navigation-alerts/NavigationAwayAlertContent";
-import { ReactComponent as CloseIcon } from "../../svgs/closeIconBold.svg";
+import TabButtons from "./TabButtons";
 import classes from "./Tabs.module.scss";
-
-const ARROW_KEY_SHIFTS = {
-  ArrowLeft: -1,
-  ArrowUp: -1,
-  ArrowRight: 1,
-  ArrowDown: 1,
-};
 
 export type TabsProps = CommonTabsProps & {
   activeTabId: string | undefined;
@@ -32,45 +25,12 @@ export const Tabs: React.FC<TabsProps> = ({
   handleUnLockDocuments,
   ...attributes
 }) => {
-  const activeTabRef = useRef<HTMLButtonElement>(null);
   const [showDocumentNavAlert, setShowDocumentNavAlert] = useState(false);
 
-  useEffect(() => {
-    activeTabRef.current?.focus();
-    activeTabRef.current?.scrollIntoView({
-      behavior: "smooth",
-      block: "nearest",
-    });
-  }, [activeTabId, items.length]);
   useLastFocus(document.querySelector("#case-details-search") as HTMLElement);
 
   const activeTabArrayPos = items.findIndex((item) => item.id === activeTabId);
   const activeTabIndex = activeTabArrayPos === -1 ? 0 : activeTabArrayPos;
-
-  const handleKeyPressOnTab: React.KeyboardEventHandler<HTMLButtonElement> = (
-    ev
-  ) => {
-    const typedKeyCode = ev.code as keyof typeof ARROW_KEY_SHIFTS;
-    const thisShift = ARROW_KEY_SHIFTS[typedKeyCode]; // -1, 1 or undefined
-    const shouldNavigate =
-      // must be a left or right key press command
-      !!thisShift &&
-      // can't go left on the first tab
-      !(activeTabIndex === 0 && thisShift === -1) &&
-      // can't go right on the last tab
-      !(activeTabIndex === items.length - 1 && thisShift === 1);
-
-    if (!shouldNavigate) {
-      return;
-    }
-
-    const nextTabIndex = activeTabIndex + ARROW_KEY_SHIFTS[typedKeyCode];
-    const nextTabId = items[nextTabIndex].id;
-    handleTabSelection(nextTabId);
-
-    // prevent awkward vertical scroll on up/down key press
-    ev.preventDefault();
-  };
 
   const handleCloseTab = () => {
     const { isDirty } = items[activeTabIndex];
@@ -103,66 +63,6 @@ export const Tabs: React.FC<TabsProps> = ({
     setShowDocumentNavAlert(false);
     localHandleClosePdf();
     handleUnLockDocuments(documentIds);
-  };
-
-  const renderTabs = () => {
-    if (!items.length) {
-      return null;
-    }
-    return (
-      <div className={`${classes.tabsWrapper}`}>
-        <ul className={`${classes.tabsList}   perma-scrollbar`} role="tablist">
-          {items.map((item, index) => {
-            const { id: itemId, label } = item;
-
-            return (
-              <li
-                className={`${
-                  activeTabIndex === index
-                    ? classes.activeTab
-                    : classes.inactiveTab
-                } ${classes.tabListItem}`}
-                key={itemId}
-                data-testid={`tab-${index}`}
-                role="presentation"
-              >
-                <button
-                  id={`tab_${index}`}
-                  aria-controls={`panel-${index}`}
-                  role="tab"
-                  className={classes.tabButton}
-                  data-testid={
-                    index === activeTabIndex ? "tab-active" : `btn-tab-${index}`
-                  }
-                  onClick={() => {
-                    if (itemId !== items[activeTabIndex].id) {
-                      handleTabSelection(itemId);
-                    }
-                  }}
-                  onKeyDown={handleKeyPressOnTab}
-                  ref={index === activeTabIndex ? activeTabRef : undefined}
-                  tabIndex={index === activeTabIndex ? 0 : -1}
-                >
-                  <span className={classes.tabLabel}> {label}</span>
-                </button>
-
-                {activeTabIndex === index && (
-                  <button
-                    className={classes.tabCloseButton}
-                    onClick={handleCloseTab}
-                    onKeyDown={handleKeyPressOnTab}
-                    data-testid="tab-remove"
-                    aria-label="close tab"
-                  >
-                    <CloseIcon />
-                  </button>
-                )}
-              </li>
-            );
-          })}
-        </ul>
-      </div>
-    );
   };
 
   const panels = items.map((item, index) => {
@@ -200,7 +100,13 @@ export const Tabs: React.FC<TabsProps> = ({
         className={`govuk-tabs ${classes.tabs} ${className || ""}`}
         {...attributes}
       >
-        {renderTabs()}
+        <TabButtons
+          items={items}
+          activeTabIndex={activeTabIndex}
+          handleTabSelection={handleTabSelection}
+          handleCloseTab={handleCloseTab}
+          handleUnLockDocuments={handleUnLockDocuments}
+        />
         {panels}
       </div>
